requirements.txt
git+https://github.com/Rapptz/discord.py
PyNaCl==1.3.0
dnspython==1.16.0
python-dotenv
